Move UI/conn ev handling to their own modules
